+++ src/app/case-studies/_ui/styles.scss
